@@ -1,31 +1,27 @@
 {
-    "name": "cdk",
-    "version": "0.1.0",
-    "bin": {
-      "cdk": "bin/cdk.js"
-    },
-    "scripts": {
-      "build": "tsc",
-      "watch": "tsc -w",
-      "test": "jest",
-      "cdk": "cdk"
-    },
-    "devDependencies": {
-      "@aws-cdk/assert": "1.123.0",
-      "@aws-cdk/aws-apigateway": "^1.124.0",
-      "@aws-cdk/aws-dynamodb": "^1.124.0",
-      "@aws-cdk/aws-lambda": "^1.124.0",
-      "@types/jest": "^26.0.10",
-      "@types/node": "10.17.27",
-      "aws-cdk": "1.123.0",
-      "jest": "^26.4.2",
-      "ts-jest": "^26.2.0",
-      "ts-node": "^9.0.0",
-      "typescript": "~3.9.7"
-    },
-    "dependencies": {
-      "@aws-cdk/core": "1.123.0",
-      "source-map-support": "^0.5.16"
-    }
+  "name": "cdk",
+  "version": "0.1.0",
+  "bin": {
+    "cdk": "bin/cdk.js"
+  },
+  "scripts": {
+    "build": "tsc",
+    "watch": "tsc -w",
+    "test": "jest",
+    "cdk": "cdk"
+  },
+  "devDependencies": {
+    "@types/jest": "^29.4.0",
+    "@types/node": "18.14.6",
+    "jest": "^29.5.0",
+    "ts-jest": "^29.0.5",
+    "aws-cdk": "2.70.0",
+    "ts-node": "^10.9.1",
+    "typescript": "~4.9.5"
+  },
+  "dependencies": {
+    "aws-cdk-lib": "2.70.0",
+    "constructs": "^10.0.0",
+    "source-map-support": "^0.5.21"
   }
-  
+}
